+++ .fallow/baseline.json
@@ -1,71 +1,28 @@
 {
-  "unused_files": [
-    "apps/webapp/src/components/DateRangePicker.tsx",
-    "apps/webapp/src/components/ThemeToggle.tsx",
-    "apps/webapp/src/components/ui/alert.tsx",
-    "apps/webapp/src/components/ui/collapsible.tsx",
-    "apps/webapp/src/components/ui/fixed-size-dialog.mdx",
-    "apps/webapp/src/components/ui/fixed-size-dialog.tsx",
-    "apps/webapp/src/modules/checklist/checklist-empty-state.tsx"
-  ],
+  "unused_files": [],
   "unused_exports": [
-    "apps/webapp/src/application/auth/RequirePermission.tsx:RequirePermission",
-    "apps/webapp/src/application/auth/index.ts:allPermissions",
-    "apps/webapp/src/application/auth/index.ts:permissions",
-    "apps/webapp/src/application/auth/index.ts:getPermissionsForRole",
-    "apps/webapp/src/application/auth/index.ts:roleDefinitions",
-    "apps/webapp/src/application/auth/index.ts:appRoles",
-    "apps/webapp/src/application/auth/index.ts:getPermissionsForUser",
-    "apps/webapp/src/application/auth/index.ts:getResolvedPermissionsForUser",
-    "apps/webapp/src/application/auth/index.ts:getRolesForUser",
-    "apps/webapp/src/application/auth/index.ts:hasPermission",
-    "apps/webapp/src/application/auth/index.ts:permissionGrantMatches",
-    "apps/webapp/src/application/auth/index.ts:unionPermissionsForRoles",
-    "apps/webapp/src/application/auth/index.ts:RequirePermission",
-    "apps/webapp/src/application/auth/permissions.ts:permissions",
     "apps/webapp/src/application/auth/resolve.ts:getRolesForUser",
     "apps/webapp/src/application/auth/resolve.ts:unionPermissionsForRoles",
     "apps/webapp/src/application/auth/resolve.ts:permissionGrantMatches",
     "apps/webapp/src/application/auth/resolve.ts:getPermissionsForUser",
     "apps/webapp/src/application/auth/resolve.ts:hasPermission",
     "apps/webapp/src/application/auth/resolve.ts:getResolvedPermissionsForUser",
-    "apps/webapp/src/application/auth/resolve.ts:appRoles",
     "apps/webapp/src/application/auth/roles.ts:systemAdminPermissions",
-    "apps/webapp/src/modules/password-protection/index.ts:generatePasswordHash",
-    "apps/webapp/src/modules/password-protection/index.ts:verifyPassword",
-    "apps/webapp/src/modules/password-protection/password-utils.ts:generatePasswordHash",
-    "services/backend/application/auth/index.ts:allPermissions",
-    "services/backend/application/auth/index.ts:permissions",
+    "apps/webapp/src/application/auth/roles.ts:roleDefinitions",
     "services/backend/application/auth/index.ts:getPermissionsForRole",
     "services/backend/application/auth/index.ts:roleDefinitions",
-    "services/backend/application/auth/index.ts:systemAdminPermissions",
-    "services/backend/application/auth/index.ts:appRoles",
-    "services/backend/application/auth/index.ts:getPermissionsForUser",
     "services/backend/application/auth/index.ts:getResolvedPermissionsForUser",
-    "services/backend/application/auth/index.ts:getRolesForUser",
-    "services/backend/application/auth/index.ts:hasPermission",
-    "services/backend/application/auth/index.ts:permissionGrantMatches",
-    "services/backend/application/auth/index.ts:unionPermissionsForRoles",
-    "services/backend/application/auth/index.ts:requirePermission",
-    "services/backend/application/auth/index.ts:requirePermissionForUser",
-    "services/backend/application/auth/permissions.ts:permissions",
     "services/backend/application/auth/requirePermission.ts:requirePermission",
     "services/backend/application/auth/requirePermission.ts:requirePermissionForUser",
     "services/backend/application/auth/resolve.ts:getRolesForUser",
     "services/backend/application/auth/resolve.ts:unionPermissionsForRoles",
     "services/backend/application/auth/resolve.ts:permissionGrantMatches",
     "services/backend/application/auth/resolve.ts:getPermissionsForUser",
-    "services/backend/application/auth/resolve.ts:appRoles",
     "services/backend/application/auth/roles.ts:systemAdminPermissions",
-    "services/backend/modules/auth/accessControl.ts:hasAccessLevel",
-    "services/backend/modules/auth/accessControl.ts:getUserAccessLevel",
-    "services/backend/modules/auth/getAuthUser.ts:getAuthUser"
+    "services/backend/application/auth/roles.ts:roleDefinitions"
   ],
   "unused_types": [
     "apps/webapp/src/application/auth/index.ts:Permission",
-    "apps/webapp/src/application/auth/index.ts:AppRole",
-    "apps/webapp/src/application/auth/index.ts:RolePermissionGrant",
-    "apps/webapp/src/application/auth/index.ts:WildcardGrant",
     "apps/webapp/src/application/auth/index.ts:UserForPermissions",
     "apps/webapp/src/application/auth/index.ts:RequirePermissionProps",
     "apps/webapp/src/modules/attendance/types.ts:AttendanceMode",
@@ -82,7 +39,7 @@
     "services/backend/modules/auth/types/AuthState.ts:GoogleUserProfile"
   ],
   "private_type_leaks": [],
-  "unused_dependencies": [],
+  "unused_dependencies": ["apps/webapp/package.json:luxon"],
   "unused_dev_dependencies": [],
   "circular_dependencies": [],
   "re_export_cycles": [],
