+++ apps/webapp/src/application/auth/index.ts
@@ -1,9 +1,4 @@
-export {
-  allPermissions,
-  permissions,
-  SYSTEM_ADMIN_ACCESS_PERMISSION,
-  type Permission,
-} from './permissions';
-export { hasPermission, type UserForPermissions } from './resolve';
+export { SYSTEM_ADMIN_ACCESS_PERMISSION, type Permission } from './permissions';
+export { type UserForPermissions } from './resolve';
 export { RequirePermission, type RequirePermissionProps } from './RequirePermission';
 export { useHasPermission } from './usePermission';
