+++ apps/webapp/src/application/auth/permissions.ts
@@ -1,7 +1,7 @@
 /**
  * Application permission registry (keep in sync with services/backend/application/auth/permissions.ts).
  */
-export const permissions = {
+const permissions = {
   'system_admin:access': {
     description:
       'Access platform system administration UI (system administrators only — not business/org admin roles)',
