+++ apps/webapp/src/modules/password-protection/index.ts
@@ -3,9 +3,5 @@
 
 export { PasswordProtect } from './PasswordProtect';
 export type { PasswordProtectConfig, PasswordProtectContextValue } from './PasswordProtectContext';
-// Context-based components
 export { PasswordProtectProvider, usePasswordProtection } from './PasswordProtectContext';
 export { PasswordProtectedConditionalRender } from './PasswordProtectedConditionalRender';
-
-// Utilities
-export { generatePasswordHash, verifyPassword } from './password-utils';
