+++ services/backend/application/auth/index.ts
@@ -1,29 +1,10 @@
-export {
-  allPermissions,
-  AUTH_PROVIDER_MANAGE_PERMISSION,
-  permissions,
-  type Permission,
-} from './permissions';
+export { AUTH_PROVIDER_MANAGE_PERMISSION, type Permission } from './permissions';
 export {
   type AppRole,
   getPermissionsForRole,
   roleDefinitions,
   type RolePermissionGrant,
-  systemAdminPermissions,
   type WildcardGrant,
 } from './roles';
-export {
-  appRoles,
-  getPermissionsForUser,
-  getResolvedPermissionsForUser,
-  getRolesForUser,
-  hasPermission,
-  permissionGrantMatches,
-  unionPermissionsForRoles,
-  type UserForPermissions,
-} from './resolve';
-export {
-  requireAuthenticatedPermission,
-  requirePermission,
-  requirePermissionForUser,
-} from './requirePermission';
+export { getResolvedPermissionsForUser, type UserForPermissions } from './resolve';
+export { requireAuthenticatedPermission } from './requirePermission';
